// src/users/entities/role.enum.ts
export enum RoleName {
  NORMAL_USER = 'NORMAL_USER',
  PASTOR = 'PASTOR',
  WORSHIP_LEADER = 'WORSHIP_LEADER',
  WORSHIP_TEAM_HEAD = 'WORSHIP_TEAM_HEAD',
  GUITARIST = 'GUITARIST',
  KEYBOARDIST = 'KEYBOARDIST',
  DRUMMER = 'DRUMMER',
  BASSIST = 'BASSIST',
  SINGER = 'SINGER',
  LEADER = 'LEADER',
  OUTREACH_WORKER = 'OUTREACH_WORKER',
  ADMIN = 'ADMIN',
  SUPER_ADMIN = 'SUPER_ADMIN',
}